% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{add_powerbi_labels}
\alias{add_powerbi_labels}
\title{Add additional columns of labels to the Power BI table}
\usage{
add_powerbi_labels(powerbi)
}
\arguments{
\item{powerbi}{The table of core data structured for Power BI.}
}
\value{
A table structured for Power BI with additional label columns.
}
\description{
Add additional columns of labels to the Power BI table
}
\keyword{internal}
