% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbenefits.R
\docType{package}
\name{clbenefits}
\alias{clbenefits}
\alias{clbenefits-package}
\title{clbenefits: Download and process a range of benefits data}
\description{
clbenefits: Download and process a range of benefits data
}
