% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{create_master}
\alias{create_master}
\title{Create the master table}
\usage{
create_master(idata)
}
\arguments{
\item{idata}{The list of dataframes returned from \code{interpolate_data}.}
}
\value{
A dataframe of key benefits data with one row per unqiue combination
  of geogpraphic area and month.
}
\description{
\code{create_master} takes the full dataset returned from
\code{interpolate_data} and combines it into a single master dataframe.
}
