% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{create_powerbi}
\alias{create_powerbi}
\title{Convert the master table into a table for Power BI}
\usage{
create_powerbi(master)
}
\arguments{
\item{master}{output table created in the pipeline.}
}
\value{
A table with the data structured for Power BI.
}
\description{
Convert the master table into a table for Power BI
}
