% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch all data from all sources}
\usage{
fetch_data(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean indicating whether the progress of the pipeline
should be printed to the console.}
}
\value{
A nested list of the results sets as dataframes.
}
\description{
\code{fetch_data} runs the full pipeline. It fetches all of the data from
each  source and stores the results as csvs in the output directory.
}
