% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrc.R
\name{fetch_hmrc_excel}
\alias{fetch_hmrc_excel}
\title{Fetch the latest hmrc tax credits data}
\usage{
fetch_hmrc_excel(url, excelname)
}
\arguments{
\item{url}{The url for hrmc geographical tax credits data.}

\item{excelname}{A filepath for the excel file download.}
}
\description{
\code{fetch_hmrc_excel} attempts to download the latest tax credits excel
file from gov.uk.
}
