% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis.R
\name{fetch_nm}
\alias{fetch_nm}
\title{Fetch all data from Nomis and return as a list of dataframes}
\usage{
fetch_nm(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean indicating whether the progress of the function
should be printed to the console.}
}
\value{
A list of the results sets as dataframes.
}
\description{
\code{fetch_nm} fetches all data from from Nomis and returns the
datasets as a list of dataframes.
}
