% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statxplore.R
\name{fetch_sx}
\alias{fetch_sx}
\title{Fetch all data from Stat-Xplore and return as a list of dataframes}
\usage{
fetch_sx(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean indicating whether the progress of the function
should be printed to the console.}
}
\value{
A list of the results sets as dataframes.
}
\description{
\code{fetch_sx} fetches all data from from Stat-Xplore and returns the
datasets as a list of dataframes.
}
