% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrc.R
\name{import_hmrc_excel}
\alias{import_hmrc_excel}
\title{Import an excel file of hmrc tax credit data and convert to a dataframe}
\usage{
import_hmrc_excel(excelname, sheet, skip, cols, date)
}
\arguments{
\item{excelname}{A filepath for an excel file of hmrc tax credits data.}

\item{sheet}{A string specifying the name of the worrksheet that contains
data on Parliamentary constituencies.}

\item{skip}{An integer specifying the number of rows to skip before the
lowest header row.}

\item{cols}{A vector of integers specifying the numbers of the columns that
contain the expected data, in the expected sequence.}

\item{date}{A date representing the date of the dataset. This is typically
the first of the month for which the dataset is published.}
}
\value{
A dataframe of the extracted tax credits data.
}
\description{
\code{import_hmrc_excel} takes the path to an excel file of hmrc tax credits
data and attempts to convert it to a csv. You must provide the function with
the information it needs to extract the correct data.
}
