% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{interpolate_data}
\alias{interpolate_data}
\title{Interpolate data from all sources (if necessary)}
\usage{
interpolate_data(fdata)
}
\arguments{
\item{fdata}{The list of dataframes returned from \code{fetch_data}.}
}
\value{
An equivalent set of dataframes which each contain monthly data.
}
\description{
\code{interpolcate_data} takes the data returned from \code{fetch_data} and
interpolates monthly data for those tables which require interpolation.
}
