% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsa.R
\name{merge_jsa}
\alias{merge_jsa}
\title{Combine historic JSA data with the Stat-Xplore JSA data and estimate the
income-based element from February 2019.}
\usage{
merge_jsa(sx_jsa, verbose = TRUE)
}
\arguments{
\item{sx_jsa}{A dataframe of JSA data returned from \code{fetch_sx}.}

\item{verbose}{A boolean indicating whether the progress of the function
should be printed to the console.}
}
\value{
Historic and current JSA data merged into a single dataframe.
}
\description{
\code{merge_jsa} combines the historic JSA figures in the input directory
with more recent figrues from Stat-Xplore. The historic data covers the
period from August 2015 to February 2019 inclusive. Data is taken from
Stat-Xplore for months since February 2019. To estimate the proportion of
JSA claimants that are income based in the Stat-Xplore data, a fixed
proportion has been calculated for February 2019 and this is reduced by
linear interpolation to the assumed end of the UC rollout period.
}
