% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Run the pipeline}
\usage{
pipeline()
}
\value{
A list containing the data, the interpolated data, and the master
  dataframe of results.
}
\description{
\code{pipeline} runs the full pipeline.
}
