% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsa.R
\name{read_historic_jsa}
\alias{read_historic_jsa}
\title{Read historic data on tax JSA and return it as a dataframe}
\usage{
read_historic_jsa(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean indicating whether the progress of the function
should be printed to the console.}
}
\value{
Historic JSA data as a dataframe.
}
\description{
\code{read_historic_jsa} reads the data on historic JSA from a csv in the
input directory and returns the data as a dataframe.
}
