% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrc.R
\name{read_hmrc}
\alias{read_hmrc}
\title{Read all hmrc data on tax credits and return it as a single dataframe}
\usage{
read_hmrc(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean indicating whether the progress of the function
should be printed to the console.}
}
\value{
A list of the results sets as dataframes.
}
\description{
\code{read_hmrc} reads all the hmrc data on tax credits stored in the
repository of csvs in the input directory and returns the data as a single
dataframe.
}
