% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbenefits.R
\docType{package}
\name{clbenefits}
\alias{clbenefits}
\title{clbenefits: Download and process a range of benefits data}
\description{
The clbenefits package implements a pipeline for dowloading and integrating
a collection of data on benefits claimed within Parliamentary
constituencies.
}
